implemented feedback task 3
