+++ src/index.js
@@ -1,10 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import MainPage from "./components/MainPage";
 import {BrowserRouter, Routes, Route} from "react-router-dom";
-import NavPage from "./components/NavPage";
 import SignIn from "./components/SignIn";
+import WeatherController from "./components/WeatherController";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
@@ -12,7 +11,7 @@ root.render(
         <Routes>
             <Route path="/" element={<App />} />
             <Route path="/login" element={<SignIn />} />
-            <Route path="/nav" element={<NavPage />} />
+            <Route path="/nav" element={<WeatherController />} />
         </Routes>
     </BrowserRouter>
 );
